chore: Moved stack codemod to v9 codemod and renamed to deprecateStack
