--- modules/codemod/lib/softDeprecate/Stack/spec/expectTransformFactory.ts
@@ -1,9 +1,0 @@
-import {runInlineTest} from 'jscodeshift/dist/testUtils';
-
-export const expectTransformFactory = (fn: Function) => (
-  input: string,
-  expected: string,
-  options?: Record<string, any>
-) => {
-  return runInlineTest(fn, options, {source: input}, expected, {parser: 'tsx'});
-};
