+++ modules/codemod/lib/v9/deprecateStack.ts
